+++ datetime/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.9-beta.1] - 2025-09-16
+
+- Fixed google sheets date bug relating to timezones.
+- Removed dependency on deprecated @epdoc/type pad function
+
 ## [1.0.9-beta.0] - 2025-09-15
 
 - Fixed regressions that occured because of partial checkin.